pyproject.toml
[tool.poetry]
name = "djangoproject"
version = "0.1.0"
description = ""
authors = ["Artsiom_PH <workartpharm@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.11.5"
django = "4.1.1"
djangorestframework = "3.14.0"
mongoengine = "0.27.0"
mypy = "1.8.0"
flake8 = "6.1.0"
djongo = "1.3.6"
pymongo = "3.12.3"
pytest = "7.4.3"
isort = "5.13.2"
black = "23.12.1"
pytest-django = "^4.7.0"
factory-boy = "^3.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 79
target-version = ["py311"]
include = '^.+\.pyi?$'
extend-exclude = '''
(
    .*\.local.*
    | .*\.idea.*
    | .*\.vscode.*
)
'''

[tool.isort]
force_single_line = true
line_length = 79
multi_line_output = 3
profile = "black"
skip = [
    ".local"
]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "src.settings"
cache_dir = ".local/pytest/"
console_output_style = "count"
testpaths = [
    "api/tests/",
]

[tool.mypy]
cache_dir = ".local/mypy/"
check_untyped_defs = true
disallow_incomplete_defs = true
disallow_untyped_calls = false
disallow_untyped_defs = false
no_implicit_optional = true
exclude = [
    ".local",
    ".github",
    ".idea",
    ".vscode",
    "venv",
    "src.settings",
]
files = ["."]
pretty = true
python_version = "3.11"
show_error_context = true
sqlite_cache = true
warn_redundant_casts = true
warn_unused_configs = true
warn_unused_ignores = true
ignore_missing_imports = true




